chore(server): use ioredis
