+++ server/libs/infra/src/index.ts
@@ -1,3 +1,4 @@
 export * from './database.config';
 export * from './infra.config';
 export * from './infra.module';
+export * from './redis-io.adapter';
