+++ server/libs/infra/src/infra.config.ts
@@ -1,16 +1,19 @@
+import { QueueName } from '@app/domain';
 import { BullModuleOptions } from '@nestjs/bull';
+import { RedisOptions } from 'ioredis';
 import { ConfigurationOptions } from 'typesense/lib/Typesense/Configuration';
-import { QueueName } from '../../domain/src';
+
+export const redisConfig: RedisOptions = {
+  host: process.env.REDIS_HOSTNAME || 'immich_redis',
+  port: parseInt(process.env.REDIS_PORT || '6379'),
+  db: parseInt(process.env.REDIS_DBINDEX || '0'),
+  password: process.env.REDIS_PASSWORD || undefined,
+  path: process.env.REDIS_SOCKET || undefined,
+};
 
 export const bullConfig: BullModuleOptions = {
   prefix: 'immich_bull',
-  redis: {
-    host: process.env.REDIS_HOSTNAME || 'immich_redis',
-    port: parseInt(process.env.REDIS_PORT || '6379'),
-    db: parseInt(process.env.REDIS_DBINDEX || '0'),
-    password: process.env.REDIS_PASSWORD || undefined,
-    path: process.env.REDIS_SOCKET || undefined,
-  },
+  redis: redisConfig,
   defaultJobOptions: {
     attempts: 3,
     removeOnComplete: true,
