+++ server/package.json
@@ -65,6 +65,7 @@
     "fluent-ffmpeg": "^2.1.2",
     "handlebars": "^4.7.7",
     "i18n-iso-countries": "^7.5.0",
+    "ioredis": "^5.3.1",
     "joi": "^17.5.0",
     "local-reverse-geocoder": "0.12.5",
     "lodash": "^4.17.21",
@@ -73,7 +74,6 @@
     "nest-commander": "^3.3.0",
     "openid-client": "^5.2.1",
     "pg": "^8.8.0",
-    "redis": "^4.5.1",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.2.0",
     "sanitize-filename": "^1.6.3",
